Added drawing as preview image
